+++ packages/component-library/stories/Popover.stories.tsx
@@ -3,8 +3,8 @@ import { Popover } from "@cultureamp/kaizen-component-library/draft"
 import { storiesOf } from "@storybook/react"
 import * as React from "react"
 
-storiesOf("Popover", module)
-  .add("Default", () => (
+storiesOf("Popover (React)", module)
+  .add("Default (Kaizen Site Demo)", () => (
     <Popover heading="Default">
       Popover body that explains something useful, is optional, and not critical
       to completing a task.